Rename utfcpp directory to utf8cpp
This is to match upstream and makes it easier for package repos to unbundle

Reference:
https://github.com/nemtrif/utfcpp
