+++ src/podofo/main/PdfCharCodeMap.cpp
@@ -8,7 +8,7 @@
 #include "PdfCharCodeMap.h"
 #include <random>
 #include <algorithm>
-#include <utfcpp/utf8.h>
+#include <utf8cpp/utf8.h>
 
 using namespace std;
 using namespace PoDoFo;
