+++ src/podofo/main/PdfDifferenceEncoding.cpp
@@ -9,7 +9,7 @@
 
 #include <algorithm>
 
-#include <utfcpp/utf8.h>
+#include <utf8cpp/utf8.h>
 #include <podofo/private/charconv_compat.h>
 
 #include "PdfArray.h"
