+++ test/common/PdfTest.cpp
@@ -8,7 +8,7 @@
 
 #include "PdfTest.h"
 
-#include <utfcpp/utf8.h>
+#include <utf8cpp/utf8.h>
 
 using namespace std;
 using namespace PoDoFo;
